Updating to install either extra modules based on kernel version
It looks like some of the rpc-openstack tests run on older images
without updating. The following changes should make a decision
to install either the linux-image-extra or linux-modules-extra
based on kernel versions.  The name changed on version 3.13.0-166
for trusy and 4.4.0-143-generic for xenial. Newton shouldn't be
installed on bionic as far as I'm aware.